requirements.txt
fastapi[all]
dotenv
uvicorn
pydantic
SQLAlchemy
requests
passlib[bcrypt]
python-multipart
python-jose[cryptography]
psycopg2
alembic
pydantic[email]
pytest
pytest-asyncio
httpx
bcrypt
alembic